feat: 支持异步消费仓库创建事件对仓库进行初始化 #3999

+++ src/backend/common/common-metadata/metadata-service/src/main/kotlin/com/tencent/bkrepo/common/metadata/service/repo/impl/RRepositoryServiceImpl.kt
@@ -103,6 +103,7 @@ import org.springframework.data.mongodb.core.query.where
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.LocalDateTime
+import java.util.Locale
 
 /**
  * 仓库服务实现类
@@ -128,6 +129,23 @@ class RRepositoryServiceImpl(
 
     override suspend fun getRepoDetail(projectId: String, name: String, type: String?): RepositoryDetail? {
         val tRepository = repositoryDao.findByNameAndType(projectId, name, type)
+        val storageCredentials = tRepository?.credentialsKey?.let { storageCredentialService.findByKey(it) }
+        return convertToDetail(tRepository, storageCredentials)
+    }
+
+    override suspend fun getRepoDetailIncludeDeleted(
+        projectId: String,
+        name: String,
+        type: String?
+    ): RepositoryDetail? {
+        val criteria = where(TRepository::projectId).isEqualTo(projectId)
+            .and(TRepository::name).isEqualTo(name)
+        val repoType = type?.uppercase(Locale.getDefault())
+        if (repoType != null && repoType != RepositoryType.NONE.name) {
+            criteria.and(TRepository::type).isEqualTo(repoType)
+        }
+
+        val tRepository = repositoryDao.findOne(Query(criteria))
         val storageCredentials = tRepository?.credentialsKey?.let { storageCredentialService.findByKey(it) }
         return convertToDetail(tRepository, storageCredentials)
     }
@@ -452,7 +470,7 @@ class RRepositoryServiceImpl(
     private suspend fun checkAndRemoveDeletedRepo(projectId: String, repoName: String, credentialsKey: String?) {
         val query = RepositoryServiceHelper.buildDeletedQuery(projectId, repoName)
         repositoryDao.findOne(query)?.let {
-            if (credentialsKey == it.credentialsKey) {
+            if (it.type != RepositoryType.DRIVE && credentialsKey == it.credentialsKey) {
                 repositoryDao.remove(query)
                 logger.info("Retrieved deleted record of Repository[$projectId/$repoName] before creating")
             } else {
